Add friends likes

+++ src/main/java/ru/company/filmorate/util/Identifier.java
@@ -6,15 +6,15 @@ import java.util.Map;
 public enum Identifier {
     INSTANCE;
 
-    private final Map<Class<?>, Integer> identifier = new HashMap<>();
+    private final Map<Class<?>, Long> identifier = new HashMap<>();
 
-    public int generate(Class<?> cl) {
+    public long generate(Class<?> cl) {
         if (identifier.containsKey(cl)) {
-            int id = identifier.get(cl);
+            long id = identifier.get(cl);
             identifier.put(cl, ++id);
             return id;
         } else {
-            identifier.put(cl, 1);
+            identifier.put(cl, 1L);
             return 1;
         }
     }
